+++ Sources/Sentry/include/HybridPublic/PrivateSentrySDKOnly.h
@@ -19,9 +19,11 @@
 @class SentryEnvelope;
 @class SentryId;
 @class SentrySessionReplayIntegration;
+@class UIView;
 
 @protocol SentryReplayBreadcrumbConverter;
 @protocol SentryViewScreenshotProvider;
+@protocol SentryRedactOptions;
 
 NS_ASSUME_NONNULL_BEGIN
 
@@ -183,6 +185,13 @@ typedef void (^SentryOnAppStartMeasurementAvailable)(
 #endif // SENTRY_UIKIT_AVAILABLE
 
 #if SENTRY_TARGET_REPLAY_SUPPORTED
+
+/**
+ * Return an instance of SentryRedactOptions with given option
+ * To be used from SentrySwiftUI, which cannot access the private
+ * `SentryRedactOptions` class.
+ */
++ (UIView *)sessionReplayMaskingOverlay:(id<SentryRedactOptions>)options;
 
 /**
  * Configure session replay with different breadcrumb converter